[Essential] Create SharedTestHelpers and move duplicated helpers
