+++ EssentialApp/EssentialAppTests/Helpers/SharedTestHelpers.swift
@@ -1,0 +1,18 @@
+import Essential_Feed
+import Foundation
+
+func uniqueFeed() -> [FeedImage] {
+    return [FeedImage(id: UUID(), description: "any", location: "any", url: URL(string: "http://any-url.com")!)]
+}
+
+func anyNSError() -> NSError {
+    return NSError(domain: "any error", code: 0)
+}
+
+func anyURL() -> URL {
+    return URL(string: "http://a-url.com")!
+}
+
+func anyData() -> Data {
+    return Data("any data".utf8)
+}
